server/requirements.txt
fastapi
google_api_python_client
google_auth_oauthlib
protobuf
pydantic
python-dotenv
Requests
starlette
uvicorn
itsdangerous 
PyJWT